Move `generate/` inside the package
