+++ litgpt/generate/adapter.py
@@ -9,13 +9,9 @@ import lightning as L
 import torch
 from lightning.fabric.plugins import BitsandbytesPrecision
 
-# support running without installing as a package
-wd = Path(__file__).parent.parent.resolve()
-sys.path.append(str(wd))
-
-from generate.base import generate
 from litgpt import Tokenizer, PromptStyle
-from litgpt.lora import GPT, Config, merge_lora_weights
+from litgpt.adapter import GPT, Config
+from litgpt.generate.base import generate
 from litgpt.prompts import load_prompt_style, has_prompt_style
 from litgpt.utils import CLI, check_valid_checkpoint_dir, get_default_supported_precision, lazy_load
 
@@ -23,32 +19,23 @@ from litgpt.utils import CLI, check_valid_checkpoint_dir, get_default_supported_
 def main(
     prompt: str = "What food do llamas eat?",
     input: str = "",
-    lora_path: Path = Path("out/lora/alpaca/lit_model_lora_finetuned.pth"),
+    adapter_path: Path = Path("out/adapter/alpaca/lit_model_adapter_finetuned.pth"),
     checkpoint_dir: Path = Path("checkpoints/stabilityai/stablelm-base-alpha-3b"),
     quantize: Optional[Literal["bnb.nf4", "bnb.nf4-dq", "bnb.fp4", "bnb.fp4-dq", "bnb.int8"]] = None,
     max_new_tokens: int = 100,
     top_k: Optional[int] = 200,
     temperature: float = 0.8,
     precision: Optional[str] = None,
-    lora_r: int = 8,
-    lora_alpha: int = 16,
-    lora_dropout: float = 0.05,
-    lora_query: bool = True,
-    lora_key: bool = False,
-    lora_value: bool = True,
-    lora_projection: bool = False,
-    lora_mlp: bool = False,
-    lora_head: bool = False,
 ) -> None:
     """Generates a response based on a given instruction and an optional input.
-    This script will only work with checkpoints from the instruction-tuned GPT-LoRA model.
-    See `litgpt/finetune/lora.py`.
+    This script will only work with checkpoints from the instruction-tuned GPT-Adapter model.
+    See `litgpt/finetune/adapter.py`.
 
     Args:
         prompt: The prompt/instruction (Alpaca style).
         input: Optional input (Alpaca style).
-        lora_path: Path to the checkpoint with trained adapter weights, which are the output of
-            `litgpt/finetune/lora.py`.
+        adapter_path: Path to the checkpoint with trained adapter weights, which are the output of
+            `litgpt/finetune/adapter.py`.
         checkpoint_dir: The path to the checkpoint folder with pretrained GPT weights.
         quantize: Whether to quantize the model and using which method:
             - bnb.nf4, bnb.nf4-dq, bnb.fp4, bnb.fp4-dq: 4-bit quantization from bitsandbytes
@@ -75,18 +62,7 @@ def main(
 
     check_valid_checkpoint_dir(checkpoint_dir)
 
-    config = Config.from_json(
-        checkpoint_dir / "lit_config.json",
-        lora_r=lora_r,
-        lora_alpha=lora_alpha,
-        lora_dropout=lora_dropout,
-        lora_query=lora_query,
-        lora_key=lora_key,
-        lora_value=lora_value,
-        lora_projection=lora_projection,
-        lora_mlp=lora_mlp,
-        lora_head=lora_head,
-    )
+    config = Config.from_json(checkpoint_dir / "lit_config.json")
 
     checkpoint_path = checkpoint_dir / "lit_model.pth"
 
@@ -112,12 +88,11 @@ def main(
 
     t0 = time.perf_counter()
     checkpoint = lazy_load(checkpoint_path)
-    lora_checkpoint = lazy_load(lora_path)
-    checkpoint.update(lora_checkpoint.get("model", lora_checkpoint))
+    adapter_checkpoint = lazy_load(adapter_path)
+    checkpoint.update(adapter_checkpoint.get("model", adapter_checkpoint))
     model.load_state_dict(checkpoint)
     fabric.print(f"Time to load the model weights: {time.perf_counter() - t0:.02f} seconds.", file=sys.stderr)
 
-    merge_lora_weights(model)
     model = fabric.setup(model)
 
     L.seed_everything(1234)
