+++ litgpt/generate/adapter_v2.py
@@ -9,20 +9,17 @@ import lightning as L
 import torch
 from lightning.fabric.plugins import BitsandbytesPrecision
 
-# support running without installing as a package
-wd = Path(__file__).parent.parent.resolve()
-sys.path.append(str(wd))
-
-from generate.base import generate
-from litgpt import GPT, Config, Tokenizer, PromptStyle
+from litgpt import Tokenizer, PromptStyle
+from litgpt.adapter_v2 import GPT, Config
+from litgpt.generate.base import generate
 from litgpt.prompts import load_prompt_style, has_prompt_style
-from litgpt.utils import CLI, check_valid_checkpoint_dir, get_default_supported_precision, load_checkpoint
+from litgpt.utils import CLI, check_valid_checkpoint_dir, get_default_supported_precision, lazy_load
 
 
 def main(
     prompt: str = "What food do llamas eat?",
     input: str = "",
-    finetuned_path: Path = Path("out/full/alpaca/lit_model_finetuned.pth"),
+    adapter_path: Path = Path("out/adapter_v2/alpaca/lit_model_adapter_finetuned.pth"),
     checkpoint_dir: Path = Path("checkpoints/stabilityai/stablelm-base-alpha-3b"),
     quantize: Optional[Literal["bnb.nf4", "bnb.nf4-dq", "bnb.fp4", "bnb.fp4-dq", "bnb.int8"]] = None,
     max_new_tokens: int = 100,
@@ -31,14 +28,14 @@ def main(
     precision: Optional[str] = None,
 ) -> None:
     """Generates a response based on a given instruction and an optional input.
-    This script will only work with checkpoints from the instruction-tuned GPT model.
-    See `litgpt/finetune/full.py`.
+    This script will only work with checkpoints from the instruction-tuned GPT-AdapterV2 model.
+    See `litgpt/finetune/adapter_v2.py`.
 
     Args:
         prompt: The prompt/instruction (Alpaca style).
         input: Optional input (Alpaca style).
-        finetuned_path: Path to the checkpoint with trained weights, which are the output of
-            `litgpt/finetune/full.py`.
+        adapter_path: Path to the checkpoint with trained adapter weights, which are the output of
+            `litgpt/finetune/adapter_v2.py`.
         checkpoint_dir: The path to the checkpoint folder with pretrained GPT weights.
         quantize: Whether to quantize the model and using which method:
             - bnb.nf4, bnb.nf4-dq, bnb.fp4, bnb.fp4-dq: 4-bit quantization from bitsandbytes
@@ -67,7 +64,7 @@ def main(
 
     config = Config.from_json(checkpoint_dir / "lit_config.json")
 
-    checkpoint_path = finetuned_path
+    checkpoint_path = checkpoint_dir / "lit_model.pth"
 
     tokenizer = Tokenizer(checkpoint_dir)
     prompt_style = load_prompt_style(checkpoint_dir) if has_prompt_style(checkpoint_dir) else PromptStyle.from_config(config)
@@ -89,11 +86,14 @@ def main(
         model.set_kv_cache(batch_size=1)
     model.eval()
 
-    model = fabric.setup(model)
-
     t0 = time.perf_counter()
-    load_checkpoint(fabric, model, checkpoint_path)
+    checkpoint = lazy_load(checkpoint_path)
+    adapter_checkpoint = lazy_load(adapter_path)
+    checkpoint.update(adapter_checkpoint.get("model", adapter_checkpoint))
+    model.load_state_dict(checkpoint)
     fabric.print(f"Time to load the model weights: {time.perf_counter() - t0:.02f} seconds.", file=sys.stderr)
+
+    model = fabric.setup(model)
 
     L.seed_everything(1234)
     t0 = time.perf_counter()
