+++ litgpt/generate/sequentially.py
@@ -17,11 +17,7 @@ from lightning.fabric.plugins import BitsandbytesPrecision
 from lightning.fabric.utilities.init import _materialize_meta_tensors
 from typing_extensions import Type
 
-# support running without installing as a package
-wd = Path(__file__).parent.parent.resolve()
-sys.path.append(str(wd))
-
-import generate.base as generate_base
+import litgpt.generate.base as generate_base
 from litgpt import GPT, Config, Tokenizer
 from litgpt.model import Block, build_mask_cache
 from litgpt.utils import CLI, check_valid_checkpoint_dir, get_default_supported_precision
