+++ tests/test_generate_lora.py
@@ -15,7 +15,7 @@ import torch
 
 @mock.patch.dict(os.environ, {"LT_ACCELERATOR": "cpu"})
 def test_main(fake_checkpoint_dir, monkeypatch, tensor_like):
-    import generate.lora as generate
+    import litgpt.generate.lora as generate
 
     config_path = fake_checkpoint_dir / "lit_config.json"
     config = {
