+++ tests/test_generate_sequentially.py
@@ -24,7 +24,7 @@ from lightning import Fabric
     ],
 )
 def test_layer_to_device(n_layer, devices, expected):
-    from generate.sequentially import layer_to_device
+    from litgpt.generate.sequentially import layer_to_device
     from litgpt.model import GPT, Block
 
     with torch.device("meta"):
@@ -40,7 +40,7 @@ def path_to_device(model):
 
 
 def test_replace_device():
-    from generate.sequentially import replace_device
+    from litgpt.generate.sequentially import replace_device
 
     class Submodule(torch.nn.Module):
         def __init__(self):
@@ -86,7 +86,7 @@ def test_replace_device():
 
 
 def _test_model_1device(accelerator):
-    from generate.sequentially import sequential
+    from litgpt.generate.sequentially import sequential
     from litgpt import GPT
 
     fabric = Fabric(accelerator=accelerator, devices=1)
@@ -157,7 +157,7 @@ def find_forward_hooks(module):
 
 @RunIf(min_cuda_gpus=2)
 def test_model_forward_hooks():
-    from generate.sequentially import sequential
+    from litgpt.generate.sequentially import sequential
     from litgpt import GPT
 
     fabric = Fabric(accelerator="cuda", devices=1)
@@ -294,9 +294,9 @@ def test_base_with_sequentially(tmp_path):
         f"--checkpoint_dir={str(checkpoint_dir)}",
     ]
     env = {"CUDA_VISIBLE_DEVICES": "0,1"}
-    base_stdout = subprocess.check_output([sys.executable, root / "generate/base.py", *args], env=env).decode()
+    base_stdout = subprocess.check_output([sys.executable, root / "litgpt/generate/base.py", *args], env=env).decode()
     sequential_stdout = subprocess.check_output(
-        [sys.executable, root / "generate/sequentially.py", *args], env=env
+        [sys.executable, root / "litgpt/generate/sequentially.py", *args], env=env
     ).decode()
 
     assert base_stdout.startswith("What food do llamas eat?")
