+++ tests/test_generate_tp.py
@@ -12,7 +12,7 @@ from test_generate_sequentially import find_forward_hooks
 
 
 def test_tensor_parallel_linear():
-    from generate.tp import tensor_parallel_linear
+    from litgpt.generate.tp import tensor_parallel_linear
 
     fabric = Mock()
     fabric.world_size = 4
@@ -82,7 +82,7 @@ def test_tensor_parallel_linear():
     ],
 )
 def test_tensor_parallel_llama(name, expected):
-    from generate.tp import tensor_parallel
+    from litgpt.generate.tp import tensor_parallel
     from litgpt import GPT
 
     fabric = Mock()
@@ -128,7 +128,7 @@ def test_tp(tmp_path):
         f"--checkpoint_dir={str(checkpoint_dir)}",
     ]
     env = {"CUDA_VISIBLE_DEVICES": "0,1"}
-    tp_stdout = subprocess.check_output([sys.executable, root / "generate/tp.py", *args], env=env).decode()
+    tp_stdout = subprocess.check_output([sys.executable, root / "litgpt/generate/tp.py", *args], env=env).decode()
 
     # there is some unaccounted randomness so cannot compare the output with that of `generate/base.py`
     assert tp_stdout.startswith("What food do llamas eat?")
