+++ tutorials/download_openllama.md
@@ -34,5 +34,5 @@ You're done! To execute the model just run:
 ```bash
 pip install sentencepiece
 
-python generate/base.py --prompt "Hello, my name is" --checkpoint_dir checkpoints/openlm-research/open_llama_3b
+python litgpt/generate/base.py --prompt "Hello, my name is" --checkpoint_dir checkpoints/openlm-research/open_llama_3b
 ```
