+++ tutorials/download_stablecode.md
@@ -40,7 +40,7 @@ You're done! To execute the model just run:
 ```bash
 pip install tokenizers
 
-python generate/base.py --prompt "Write in Python a softmax function. Be concise." --checkpoint_dir checkpoints/$repo_id
+python litgpt/generate/base.py --prompt "Write in Python a softmax function. Be concise." --checkpoint_dir checkpoints/$repo_id
 ```
 
 Or you can run the model in an interactive mode:
