+++ tutorials/finetune_lora.md
@@ -78,7 +78,7 @@ For additional benchmarks and resource requirements, please see the [Resource Ta
 You can test the finetuned model with your own instructions by running:
 
 ```bash
-python generate/lora.py \
+python litgpt/generate/lora.py \
   --prompt "Recommend a movie to watch on the weekend."
 ```
 
@@ -148,10 +148,10 @@ python scripts/merge_lora.py \
   --checkpoint_dir "out/lora/stablelm-base-alpha-3b/final"
 ```
 
-After merging, we can use the `generate/base.py` or `chat/base.py` file for inference using the new checkpoint file. 
+After merging, we can use the `litgpt/generate/base.py` or `chat/base.py` file for inference using the new checkpoint file. 
 
 ```bash
-python generate/base.py \
+python litgpt/generate/base.py \
   --checkpoint_dir "out/lora/stablelm-base-alpha-3b/final"
 ```
